requirements.txt
alpaca-trade-api==3.0.2
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
tensorflow==2.15.0
discord.py==2.3.2
pytz==2023.3
aiohttp==3.9.1
giotto-tda==0.6.0
ripser==0.6.4
persim==0.3.1
ta==0.11.0
